Scripts: Add test-e2e script to wp-scripts (#12437)
* Scripts: Add test-e2e script

* Docs: Add docs for test-e2e script added to scripts packages

* Update package-lock.json

* Refactor scripts based on feedback shared in the review

* More fixes to wp-scripts test-e2e

* Address feedback shared in the review

* Fix runInBand flag usage
