+++ packages/babel-preset-default/package.json
@@ -23,6 +23,7 @@
 	},
 	"main": "index.js",
 	"dependencies": {
+		"@babel/core": "^7.0.0",
 		"@babel/plugin-proposal-async-generator-functions": "^7.0.0",
 		"@babel/plugin-proposal-object-rest-spread": "^7.0.0",
 		"@babel/plugin-transform-react-jsx": "^7.0.0",
