+++ packages/scripts/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### New Features
 
 - Added support for `check-engines` script ([#12721](https://github.com/WordPress/gutenberg/pull/12721))
+- Added support for `test-e2e` script ([#12437](https://github.com/WordPress/gutenberg/pull/12437)).
 - Update default config provided for `lint-js` script ([#12845](https://github.com/WordPress/gutenberg/pull/12845)).
 
 ## 2.4.4 (2018-11-20)
