+++ packages/scripts/scripts/test-unit-jest.js
@@ -18,22 +18,13 @@ const jest = require( 'jest' );
  * Internal dependencies
  */
 const {
+	fromConfigRoot,
 	getCliArgs,
-	hasCliArg,
-	hasProjectFile,
-	hasPackageProp,
+	hasJestConfig,
 } = require( '../utils' );
 
-const args = getCliArgs();
-
-const hasJestConfig = hasCliArg( '-c' ) ||
-	hasCliArg( '--config' ) ||
-	hasProjectFile( 'jest.config.js' ) ||
-	hasProjectFile( 'jest.config.json' ) ||
-	hasPackageProp( 'jest' );
-
-const config = ! hasJestConfig ?
-	[ '--config', JSON.stringify( require( '../config/jest.config' ) ) ] :
+const config = ! hasJestConfig() ?
+	[ '--config', JSON.stringify( require( fromConfigRoot( 'jest-unit.config.js' ) ) ) ] :
 	[];
 
-jest.run( [ ...config, ...args ] );
+jest.run( [ ...config, ...getCliArgs() ] );
